Add proof of concept CoberturaTarHandler
This XML handler takes a tarball of Cobertura XML files and adds
coverage information from them. The tarball must also include the
source files listed in the Cobertura files.

This was just a proof of concept I came up with before I ended up
going in a different direction. Hopefully someone can figure out
what to do with this and finish it off.